Move TestIndexingSimplification to its own file
[ghstack-poisoned]
